Caesar Cypher
This is an exercise from the course "100 Days of Code: The Complete Python Pro Bootcamp"
Basically it cyphers any messages according to the old Caesar's encrypting method. It shifts the letters according to the value the user enters as shift. So anyone who knows the numbers of shifts can decypher the message with the same program, by typing "decode" instead of "encode".